checkpoint

+++ lib/shared/_responsive/responsive.dart
@@ -26,9 +26,9 @@ class Responsive extends StatelessWidget {
   Widget build(BuildContext context) {
     return LayoutBuilder(
       builder: (context, constraints) {
-        if (constraints.maxWidth >= 1200) {
+        if (constraints.maxWidth.toInt() >= 1200) {
           return desktop;
-        } else if (constraints.maxWidth >= 800) {
+        } else if (constraints.maxWidth.toInt() >= 800) {
           return tablet ?? mobile;
         } else {
           return mobile;
